# app/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://192.168.1.98:8280/api",
            "description": "Сервер"
        },
        {
            "url": ""
        }
    ],
    "paths": {
...
        "/cars": {
            "get": {
                "tags": [
                    "Новый автомобиль"
                ],
                "summary": "Список новых автомобилей",
                "description": "Список новых автомобилейв",
                "operationId": "carsList",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CarFilter"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            },
            "post": {
                "tags": [
                    "Новый автомобиль"
                ],
                "summary": "Создать новый автомобиль",
                "description": "Создать новый автомобиль",
                "operationId": "storeCar",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CarCreateRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/cars/{id}": {
            "get": {
                "tags": [
                    "Новый автомобиль"
                ],
                "summary": "Открыть карточку нового автомобиля",
                "description": "Открыть карточку нового автомобиля",
                "operationId": "getCar",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Идентификатор автомобиля",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Новый автомобиль"
                ],
                "summary": "Удалить автомобиль",
                "description": "Удалить автомобиль",
                "operationId": "deleteCar",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Идентификатор автомобиля",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            },
            "patch": {
                "tags": [
                    "Новый автомобиль"
                ],
                "summary": "Изменить новый автомобиль",
                "description": "Изменить новый автомобиль",
                "operationId": "updateCar",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Идентификатор машины",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CarCreateRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/cars/overprice/{id}": {
            "get": {
                "tags": [
                    "Новый автомобиль"
                ],
                "summary": "Получить стоимость переоценки автомобиля (воздух)",
                "description": "Получить стоимость переоценки автомобиля (воздух)",
                "operationId": "getCarOverPrice",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Идентификатор автомобиля",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            },
            "post": {
                "tags": [
                    "Новый автомобиль"
                ],
                "summary": "Добавить/Изменить стоимость переоценки автомобиля (воздух)",
                "description": "Добавить/Изменить стоимость переоценки автомобиля (воздух)",
                "operationId": "storeCarOverPrice",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Идентификатор автомобиля",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "price",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "price"
                                ],
                                "properties": {
                                    "price": {
                                        "type": "integer",
                                        "format": "integer",
                                        "example": "10000"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/cars/{id}/history": {
            "get": {
                "tags": [
                    "Новый автомобиль"
                ],
                "summary": "Получить историю автомобиля",
                "description": "Получить историю автомобиля",
                "operationId": "getCarHistory",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Идентификатор автомобиля",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/cars/{id}/tuning": {
            "get": {
                "tags": [
                    "Новый автомобиль"
                ],
                "summary": "Получить установленный тюнинг на автомобиле",
                "description": "Получить установленный тюнинг на автомобиле",
                "operationId": "getCarTuning",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Идентификатор автомобиля",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "cars/restore/{id}": {
            "patch": {
                "tags": [
                    "Востановить автомобиль"
                ],
                "summary": "Востановить автомобиль",
                "description": "Востановить автомобиль",
                "operationId": "restoreCar",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Идентификатор автомобиля",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "CarCreateRequest": {
                "required": [
                    "mark_id",
                    "brand_id",
                    "complectation_id",
                    "color_id",
                    "year",
                    "trade_marker_id",
                    "provider_id"
                ]
            },
...
        },
...
    },
...
}